URL and MySQL Injection Protection

+++ admin/includes/view_all_users.php
@@ -33,32 +33,10 @@
                 $output  = "<td>{$user_id}</td>";
                 $output .= "<td>{$username}</td>";
                 $output .= "<td>{$user_firstname}</td>";
-
-                // $query = "SELECT * FROM categories WHERE cat_id = {$post_category_id}";
-                // $select_categories_id = mysqli_query( $connection, $query );
-
-                // while ( $row = mysqli_fetch_assoc( $select_categories_id ) ) {
-                //     $cat_id = $row['cat_id'];
-                //     $cat_title = $row['cat_title'];
-                
-                //     $output .= "<td>{$cat_title}</td>";
-                // }
-
                 $output .= "<td>{$user_lastname}</td>";
                 $output .= "<td>{$user_email}</td>";
                 $output .= "<td>{$user_role}</td>";
-
-                //  $query = "SELECT * FROM posts WHERE post_id=$comment_post_id";
-                //  $select_post_id_query = mysqli_query( $connection, $query );
-
-                //  while ( $row = mysqli_fetch_assoc( $select_post_id_query ) ) {
-                //  $post_id = $row['post_id'];
-                //  $post_title = $row['post_title'];
-
-				// 	$output .= "<td><a href='../post.php?p_id=$post_id'>{$post_title}</a></td>";
-				// }
-				
-                $output .= "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
+				$output .= "<td><a href='users.php?change_to_admin={$user_id}'>Admin</a></td>";
                 $output .= "<td><a href='users.php?change_to_sub={$user_id}'>Subscriber</a></td>";
                 $output .= "<td><a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a></td>";
                 $output .= "<td><a href='users.php?delete={$user_id}'>Delete</a></td>";
@@ -95,11 +73,20 @@
 
     // Delete user from the database
     if ( isset( $_GET['delete'] ) )  {
-        $the_user_id = $_GET['delete']; 
 
-        $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-        $delete_user_query = mysqli_query( $connection, $query );
-        header( "Location: users.php" );
+        if ( isset( $_SESSION['user_role'] ) ) {
+
+            if ( $_SESSION['user_role'] == 'admin' ) {
+
+                $the_user_id = mysqli_real_escape_string( $connection,  $_GET['delete'] );
+
+                $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+                $delete_user_query = mysqli_query( $connection, $query );
+                header( "Location: users.php" );
+            }
+
+        }
+
     }
     
 ?>
